services/ai-agent/requirements.txt
transformers
torch
scikit-learn
kafka-python